got rid of test_async_callable_sink.py -- Check for weird __call__ edge case -- formatted src and tets with black -l 100
