+++ loguru/_logger.py
@@ -98,6 +98,7 @@ from ._locks_machinery import create_logger_lock
 from ._recattrs import RecordException, RecordFile, RecordLevel, RecordProcess, RecordThread
 from ._simple_sinks import AsyncSink, CallableSink, StandardSink, StreamSink
 
+
 if sys.version_info >= (3, 6):
     from os import PathLike
 else:
@@ -109,7 +110,6 @@ elif sys.version_info >= (3, 5, 3):
     from aiocontextvars import ContextVar
 else:
     from contextvars import ContextVar
-
 
 Level = namedtuple("Level", ["name", "no", "color", "icon"])
 
@@ -790,7 +790,9 @@ class Logger:
             encoding = getattr(sink, "encoding", None)
             terminator = ""
             exception_prefix = "\n"
-        elif inspect.iscoroutinefunction(sink) or (callable(sink) and inspect.iscoroutinefunction(sink.__call__)):
+        elif inspect.iscoroutinefunction(sink) or inspect.iscoroutinefunction(
+            getattr(sink, "__call__", None)
+        ):
             name = getattr(sink, "__name__", None) or repr(sink)
 
             if colorize is None:
@@ -806,7 +808,8 @@ class Logger:
             if enqueue and loop is None:
                 loop = asyncio.get_event_loop()
 
-            wrapped_sink = AsyncSink(sink, loop, error_interceptor)
+            coro = sink if inspect.iscoroutinefunction(sink) else sink.__call__
+            wrapped_sink = AsyncSink(coro, loop, error_interceptor)
             encoding = "utf8"
             terminator = "\n"
             exception_prefix = ""
@@ -1999,7 +2002,6 @@ class Logger:
     @staticmethod
     @functools.lru_cache(maxsize=32)
     def _dynamic_level(level):
-
         if isinstance(level, str):
             return (level, None)
