+++ tests/exceptions/source/backtrace/nested_explicit_catch_up.py
@@ -14,6 +14,7 @@ def foo():
 def bar():
     1 / 0
 
+
 try:
     foo()
 except ZeroDivisionError:
