+++ tests/exceptions/source/diagnose/attributes.py
@@ -6,7 +6,6 @@ logger.add(sys.stderr, format="", colorize=True, backtrace=False, diagnose=True)
 
 
 class Obj:
-
     @property
     def forbidden(self):
         raise RuntimeError
@@ -18,7 +17,8 @@ a.b = "123"
 
 def foo():
     x = None
-    ... + 1 + bar(a).b + a.forbidden + a.nope.a + x.__bool__ or a. b . isdigit() and .3 + ...
+    ... + 1 + bar(a).b + a.forbidden + a.nope.a + x.__bool__ or a.b.isdigit() and 0.3 + ...
+
 
 try:
     foo()
