+++ tests/exceptions/source/diagnose/global_variable.py
@@ -8,9 +8,11 @@ logger.add(sys.stderr, format="", colorize=True, backtrace=False, diagnose=True)
 foo = True
 bar = False
 
+
 def func():
     foo = None
     return 1 / 0 + foo + bar + False
+
 
 try:
     func()
