+++ tests/exceptions/source/diagnose/source_strings.py
@@ -8,6 +8,6 @@ logger.add(sys.stderr, format="", colorize=True, backtrace=False, diagnose=True)
 a = b = 0
 
 try:
-    a + b"prefix" + 'single' + """triple""" + 1 + b
+    a + b"prefix" + "single" + """triple""" + 1 + b
 except TypeError:
     logger.exception("")
