+++ tests/exceptions/source/diagnose/unprintable_object.py
@@ -4,10 +4,11 @@ from loguru import logger
 logger.remove()
 logger.add(sys.stderr, format="", colorize=True, backtrace=False, diagnose=True)
 
-class Object:
 
+class Object:
     def __repr__(self):
         raise RuntimeError("No way!")
+
 
 try:
     obj = Object()
